quequierocortarlamrcorneta

+++ VITIS_HLS_SRC/specs.h
@@ -2,7 +2,8 @@
 #ifndef SPECS_H
 #define SPECS_H
 
-typedef int T_in;
+//typedef int T_in;
+typedef float T_in;
 
 #define LENGTH 1024           /* Largo de vectores para 1024 palabras */
 //#define LENGTH 128          /* Largo de vectores para 128 palabras */
